fix: cart 컴포넌트 내부에 존재하던 calcCart 함수를 CartStore 내부로 분리 후 해당 함수에 파악을 위한 주석 추가
