+++ src/basic/__tests__/basic.test.js
@@ -21,7 +21,9 @@ describe('basic test', () => {
     });
 
     beforeEach(() => {
-      vi.useRealTimers();
+      vi.useFakeTimers();
+      const mockDate = new Date('2025-1-6');
+      vi.setSystemTime(mockDate);
       vi.spyOn(window, 'alert').mockImplementation(() => {});
     });
